WhatBytes/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
# psycopg
python-dotenv
